requirements/dev-py37.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev-py37.txt requirements/dev.in
#
appdirs==1.4.4            # via black, virtualenv
asgiref==3.2.10           # via django
astroid==2.4.2            # via pylint
autopep8==1.5.4           # via -r requirements/dev.in
babel==2.8.0              # via django-prices, enmerkar, prices
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements/dev.in
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black, pip-tools
coverage==5.3             # via -r requirements/dev.in
decorator==4.4.2          # via ipython
distlib==0.3.1            # via virtualenv
django-braces==1.14.0     # via django-model-reviews, django-vega-admin
django-contrib-comments==1.9.2  # via django-model-reviews
django-crispy-forms==1.9.2  # via django-vega-admin, small-small-hr
django-filter==2.3.0      # via django-vega-admin
django-js-asset==1.2.2    # via django-mptt
django-model-reviews==1.1.0  # via small-small-hr
django-mptt==0.11.0       # via small-small-hr
django-phonenumber-field==5.0.0  # via django-tiny-erp, small-small-hr
django-prices==2.2.0      # via django-tiny-erp
django-private-storage==2.2.2  # via small-small-hr
django-tables2==2.3.1     # via django-vega-admin
django-vega-admin==0.0.19  # via django-tiny-erp
django==3.1.1             # via django-braces, django-filter, django-model-reviews, django-mptt, django-phonenumber-field, django-prices, django-tables2, django-tiny-erp, django-vega-admin, djangorestframework, enmerkar, model-bakery, small-small-hr
djangorestframework==3.11.1  # via small-small-hr
enmerkar==0.7.1           # via django-prices
fastdiff==0.2.0           # via snapshottest
filelock==3.0.12          # via tox, virtualenv
flake8==3.8.3             # via -r requirements/dev.in
identify==1.5.3           # via pre-commit
importlib-metadata==1.7.0  # via flake8, pluggy, pre-commit, tox, virtualenv
ipdb==0.13.3              # via -r requirements/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.18.1           # via ipdb
isort==5.5.2              # via -r requirements/dev.in, pylint
jedi==0.17.2              # via ipython
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via flake8, pylint
model-bakery==1.1.1       # via -r requirements/dev.in
mypy-extensions==0.4.3    # via black, mypy
mypy==0.782               # via -r requirements/dev.in
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via tox
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pep8==1.7.1               # via -r requirements/dev.in
pexpect==4.8.0            # via ipython
phonenumbers==8.12.9      # via django-tiny-erp
phonenumberslite==8.12.9  # via small-small-hr
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via small-small-hr
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via tox
pre-commit==2.7.1         # via -r requirements/dev.in
prices==1.1.0             # via django-prices
prompt-toolkit==3.0.7     # via ipython
psycopg2-binary==2.8.6    # via -r requirements/dev.in, small-small-hr
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via tox
pycodestyle==2.6.0        # via -r requirements/dev.in, autopep8, flake8
pydocstyle==5.1.1         # via -r requirements/dev.in
pyflakes==2.2.0           # via flake8
pygments==2.7.1           # via ipython
pylint-django==2.3.0      # via -r requirements/dev.in
pylint-plugin-utils==0.6  # via pylint-django
pylint==2.6.0             # via -r requirements/dev.in, pylint-django, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via babel, django
pyyaml==5.3.1             # via django-vega-admin, pre-commit
regex==2020.7.14          # via black
six==1.15.0               # via astroid, django-braces, packaging, pip-tools, snapshottest, tox, virtualenv
small_small_hr==0.2.2     # via django-tiny-erp
snapshottest==0.5.1       # via -r requirements/dev.in
snowballstemmer==2.0.0    # via pydocstyle
sorl-thumbnail==12.6.3    # via small-small-hr
sqlparse==0.3.1           # via django
tablib==2.0.0             # via django-vega-admin
termcolor==1.1.0          # via snapshottest
toml==0.10.1              # via autopep8, black, pre-commit, pylint, tox
tox==3.20.0               # via -r requirements/dev.in
traitlets==5.0.4          # via ipython
typed-ast==1.4.1          # via astroid, black, mypy
typing-extensions==3.7.4.3  # via black, mypy
urllib3==1.25.10          # via -r requirements/dev.in
virtualenv==20.0.31       # via pre-commit, tox
voluptuous==0.11.7        # via small-small-hr
wasmer==0.4.1             # via fastdiff
wcwidth==0.2.5            # via prompt-toolkit
wrapt==1.12.1             # via astroid
yapf==0.30.0              # via -r requirements/dev.in
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
